fix button styles / layout.
