+++ src/app/(main)/my-voyage/[teamId]/tech-stack/components/AddVoteBtn.tsx
@@ -1,23 +1,9 @@
-import { EllipsisVerticalIcon } from "@heroicons/react/24/solid";
 import Button from "@/components/Button";
 
-interface AddVoteBtnProps {
-  id: number;
-  openMenu: (value: number) => void;
-}
-
-export default function AddVoteBtn({ id, openMenu }: AddVoteBtnProps) {
-  const handleClick = () => {
-    openMenu(id);
-  };
-
+export default function AddVoteBtn() {
   return (
-    <div className="flex col-span-2">
-      <EllipsisVerticalIcon
-        className="w-1/6 h-1/6 hover:cursor-pointer"
-        onClick={handleClick}
-      />
-      <Button variant="primary" size="xs" className="rounded-3xl w-5/6">
+    <div className="col-span-2 flex justify-end">
+      <Button variant="primary" size="xs" className="rounded-3xl">
         Add Vote
       </Button>
     </div>
