+++ src/app/(main)/my-voyage/[teamId]/tech-stack/components/RemoveVoteBtn.tsx
@@ -1,13 +1,29 @@
+import { EllipsisVerticalIcon } from "@heroicons/react/24/solid";
 import Button from "@/components/Button";
 
-export default function RemoveVoteBtn() {
+interface RemoveVoteBtnProps {
+  id: number;
+  openMenu: (value: number) => void;
+}
+
+export default function RemoveVoteBtn({ id, openMenu }: RemoveVoteBtnProps) {
+  const handleClick = () => {
+    openMenu(id);
+  };
+
   return (
-    <Button
-      variant="error"
-      size="xs"
-      className="w-11/12 rounded-3xl col-span-2 justify-self-end"
-    >
-      Remove Vote
-    </Button>
+    <div className="flex justify-end col-span-2">
+      <EllipsisVerticalIcon
+        className="w-1/6 h-1/6 hover:cursor-pointer"
+        onClick={handleClick}
+      />
+      <Button
+        variant="error"
+        size="xs"
+        className="rounded-3xl col-span-2 justify-self-end"
+      >
+        Remove Vote
+      </Button>
+    </div>
   );
 }
